requirements.txt
python-nmap
requests
ttkbootstrap